feat: Parses JSON and checks for requestToken
Parsed results are tested to see if the JSON can be parsed. If the
JSON can be parsed check for valid status code, check that request
token is return. If all statements pass the checks, set AppDelegate
request token from nil to the request token value. Also makes a call
to loginWithToken request and passes the token to this function.